Revert "EOS-16481: Added new resource agent (#279)"
This reverts commit 2f79e16fae760fa58c6ee9a62680fedf93049623.
