+++ ha/resource/hw_comp_ra.py
@@ -28,6 +28,6 @@ import pathlib
 
 if __name__ == '__main__':
     sys.path.append(os.path.join(os.path.dirname(pathlib.Path(__file__)), '..', '..'))
-    from ha.resource import alert_monitor_resource_agent
+    from ha.resource import resource_agent
     action = sys.argv[1] if len(sys.argv) > 1 else ""
-    sys.exit(alert_monitor_resource_agent.main(alert_monitor_resource_agent.HardwareResourceAgent, action))
+    sys.exit(resource_agent.main(resource_agent.HardwareResourceAgent, action))
