+++ jenkins/pyinstaller/v2/pyinstaller-cortx-ha.spec
@@ -16,6 +16,7 @@
 # about this software or licensing, please email opensource@seagate.com or
 # cortx-questions@seagate.com.
 
+
 import sys
 import os
 import re
@@ -26,79 +27,115 @@ block_cipher = None
 ha_path="<HA_PATH>"
 
 # Analysis
+hw_comp_ra = Analysis([ha_path + '/ha/resource/hw_comp_ra.py'],
+             pathex=[ha_path],
+             binaries=[],
+             datas=[],
+             hiddenimports=[],
+             hookspath=[],
+             runtime_hooks=[],
+             excludes=['numpy', 'matplotlib'],
+             win_no_prefer_redirects=False,
+             win_private_assemblies=False,
+             cipher=block_cipher,
+             noarchive=False)
+
+iem_comp_ra = Analysis([ha_path + '/ha/resource/iem_comp_ra.py'],
+             pathex=[ha_path],
+             binaries=[],
+             datas=[],
+             hiddenimports=[],
+             hookspath=[],
+             runtime_hooks=[],
+             excludes=['numpy', 'matplotlib'],
+             win_no_prefer_redirects=False,
+             win_private_assemblies=False,
+             cipher=block_cipher,
+             noarchive=False)
+
 cortxha =  Analysis([ha_path + '/ha/cli/cortxha.py'],
-        pathex=[ha_path],
-        binaries=[],
-        datas=[],
-        hiddenimports=[],
-        hookspath=[],
-        runtime_hooks=[],
-        excludes=['numpy', 'matplotlib'],
-        win_no_prefer_redirects=False,
-        win_private_assemblies=False,
-        cipher=block_cipher,
-        noarchive=False)
+             pathex=[ha_path],
+             binaries=[],
+             datas=[],
+             hiddenimports=[],
+             hookspath=[],
+             runtime_hooks=[],
+             excludes=['numpy', 'matplotlib'],
+             win_no_prefer_redirects=False,
+             win_private_assemblies=False,
+             cipher=block_cipher,
+             noarchive=False)
 
-dynamic_fid_service_ra =  Analysis([ha_path + '/ha/resource/dynamic_fid_service_ra.py'],
-        pathex=[ha_path],
-        binaries=[],
-        datas=[],
-        hiddenimports=[],
-        hookspath=[],
-        runtime_hooks=[],
-        excludes=['numpy', 'matplotlib'],
-        win_no_prefer_redirects=False,
-        win_private_assemblies=False,
-        cipher=block_cipher,
-        noarchive=False)
 
-MERGE((cortxha, 'cortxha', 'cortxha'),
-        (dynamic_fid_service_ra, 'dynamic_fid_service_ra', 'dynamic_fid_service_ra'))
+MERGE( (hw_comp_ra, 'hw_comp_ra', 'hw_comp_ra'),
+        (iem_comp_ra, 'iem_comp_ra', 'iem_comp_ra'),
+        (cortxha, 'cortxha', 'cortxha'))
+
+# hw_comp_ra
+hw_comp_ra_pyz = PYZ(hw_comp_ra.pure, hw_comp_ra.zipped_data,
+             cipher=block_cipher)
+
+hw_comp_ra_exe = EXE(hw_comp_ra_pyz,
+          hw_comp_ra.scripts,
+          [],
+          exclude_binaries=True,
+          name='hw_comp_ra',
+          debug=False,
+          bootloader_ignore_signals=False,
+          strip=False,
+          upx=True,
+          console=True )
+
+# iem_comp_ra
+iem_comp_ra_pyz = PYZ(iem_comp_ra.pure, iem_comp_ra.zipped_data,
+             cipher=block_cipher)
+
+iem_comp_ra_exe = EXE(iem_comp_ra_pyz,
+          iem_comp_ra.scripts,
+          [],
+          exclude_binaries=True,
+          name='iem_comp_ra',
+          debug=False,
+          bootloader_ignore_signals=False,
+          strip=False,
+          upx=True,
+          console=True )
 
 # cortxha
 cortxha_pyz = PYZ(cortxha.pure, cortxha.zipped_data,
-        cipher=block_cipher)
+             cipher=block_cipher)
 
 cortxha_exe = EXE(cortxha_pyz,
-        cortxha.scripts,
-        [],
-        exclude_binaries=True,
-        name='cortxha',
-        debug=False,
-        bootloader_ignore_signals=False,
-        strip=False,
-        upx=True,
-        console=True )
-
-# dynamic_fid_service_ra
-dynamic_fid_service_ra_pyz = PYZ(dynamic_fid_service_ra.pure, dynamic_fid_service_ra.zipped_data,
-        cipher=block_cipher)
-
-dynamic_fid_service_ra_exe = EXE(dynamic_fid_service_ra_pyz,
-        dynamic_fid_service_ra.scripts,
-        [],
-        exclude_binaries=True,
-        name='dynamic_fid_service_ra',
-        debug=False,
-        bootloader_ignore_signals=False,
-        strip=False,
-        upx=True,
-        console=True )
+          cortxha.scripts,
+          [],
+          exclude_binaries=True,
+          name='cortxha',
+          debug=False,
+          bootloader_ignore_signals=False,
+          strip=False,
+          upx=True,
+          console=True )
 
 coll = COLLECT(
-        # cortxha
-        cortxha_exe,
-        cortxha.binaries,
-        cortxha.zipfiles,
-        cortxha.datas,
+               # hw_comp_ra
+               hw_comp_ra_exe,
+               hw_comp_ra.binaries,
+               hw_comp_ra.zipfiles,
+               hw_comp_ra.datas,
 
-        # dynamic_fid_service_ra
-        dynamic_fid_service_ra_exe,
-        dynamic_fid_service_ra.binaries,
-        dynamic_fid_service_ra.zipfiles,
-        dynamic_fid_service_ra.datas,
+               # iem_comp_ra
+               iem_comp_ra_exe,
+               iem_comp_ra.binaries,
+               iem_comp_ra.zipfiles,
+               iem_comp_ra.datas,
 
-        strip=False,
-        upx=True,
-        upx_exclude=[],
-        name='lib')
+               # cortxha
+               cortxha_exe,
+               cortxha.binaries,
+               cortxha.zipfiles,
+               cortxha.datas,
+
+               strip=False,
+               upx=True,
+               upx_exclude=[],
+               name='lib')
